pyproject.toml
[tool.poetry]
name = "ded-healthbot"
version = "0.1.0"
description = "Telegram-бот мониторинга health-эндпоинтов сервисов"
authors = ["GasKhan <blck6rd@proton.me>"]
readme = "README.md"
packages = [{ include = "healthbot" }]

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.13"
aiohttp = "^3.10"
pydantic = "^2.10"
pydantic-settings = "^2.6"
pyyaml = "^6.0"
aiofiles = "^24.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 78
target-version = "py313"
